add part permissions & and solve some of issues in admin panel

+++ resources/views/admin/pages/categories/get-data.blade.php
@@ -111,16 +111,32 @@
                                         Actions <i class="bx bx-dots-vertical-rounded"></i>
                                         </button>
                                         <div class="dropdown-menu">
-                                            <a class="dropdown-item" href="javascript:void(0);"
-                                                wire:click="$dispatch('categoryUpdate', { id: {{ $value->id }} })"
-                                            >
-                                                <i class="bx bx-edit-alt me-1"></i> Edit
-                                            </a>
-                                            <a class="dropdown-item" href="#"
-                                                wire:click.prevent="$dispatch('categoryDelete', { id: {{ $value->id }} })"
-                                            >
-                                                <i class="bx bx-trash me-1"></i> Delete
-                                            </a>
+
+                                            @if(admin()->can('category-read'))
+                                                <a class="dropdown-item" href="javascript:void(0);"
+                                                    wire:click="$dispatch('categoryShow', { id: {{ $value->id }} })"
+                                                >
+                                                    <i class="bx bx-show me-1"></i> Show
+                                                </a>                                    
+                                            @endif
+
+                                            @if(admin()->can('category-edit'))
+                                                <a class="dropdown-item" href="javascript:void(0);"
+                                                    wire:click="$dispatch('categoryUpdate', { id: {{ $value->id }} })"
+                                                >
+                                                    <i class="bx bx-edit-alt me-1"></i> Edit
+                                                </a>                                    
+                                            @endif
+                                
+                                            @if(admin()->can('category-delete'))
+                                                <a class="dropdown-item" href="#"
+                                                    wire:click.prevent="$dispatch('categoryDelete', { id: {{ $value->id }} })"
+                                                >
+                                                    <i class="bx bx-trash me-1"></i> Delete
+                                                </a>
+                                            @endif
+
+
 
 
                                         </div>
