+++ resources/views/admin/pages/keepers/get-data.blade.php
@@ -78,36 +78,47 @@
                                                 Actions <i class="bx bx-dots-vertical-rounded"></i>
                                                 </button>
                                                 <div class="dropdown-menu">
-                                                    <a class="dropdown-item" href="javascript:void(0);"
-                                                        wire:click.prevent="$dispatch('keeperShow', { id: {{ $value->id }} })"
-                                                    >
-                                                        <i class="bx bx-show me-1"></i> show
-                                                    </a>
-                                                    <a class="dropdown-item" href="javascript:void(0);"
-                                                        wire:click.prevent="$dispatch('keeperAssignRole', { id: {{ $value->id }} })"
-                                                    >
-                                                        <i class="bx bx-check me-1"></i> Assign Role
-                                                    </a>
 
-                                                    <a class="dropdown-item" href="javascript:void(0);"
-                                                        wire:click.prevent="$dispatch('keeperVerify', { id: {{ $value->id }} })"
-                                                    >
-                                                        <i class="bx bx-envelope me-1"></i> {{ $value->email_verified_at ? 'Not Verify' : 'Verify' }}
-                                                    </a>
+                                                    @if(admin()->can('keeper-read'))
+                                                        <a class="dropdown-item" href="javascript:void(0);"
+                                                            wire:click.prevent="$dispatch('keeperShow', { id: {{ $value->id }} })"
+                                                        >
+                                                            <i class="bx bx-show me-1"></i> show
+                                                        </a>
+                                                    @endif
 
+                                                    @if(admin()->can('keeper-assign-role'))
+                                                        <a class="dropdown-item" href="javascript:void(0);"
+                                                            wire:click.prevent="$dispatch('keeperAssignRole', { id: {{ $value->id }} })"
+                                                        >
+                                                            <i class="bx bx-check me-1"></i> Assign Role
+                                                        </a>
+                                                    @endif
 
-                                                    <a class="dropdown-item" href="javascript:void(0);"
-                                                        wire:click.prevent="$dispatch('keeperUpdate', { id: {{ $value->id }} })"
-                                                    >
-                                                        <i class="bx bx-edit-alt me-1"></i> Edit
-                                                    </a>
-                        
-                                                    <a class="dropdown-item" href="javascript:void(0);"
-                                                        wire:click.prevent="$dispatch('keeperDelete', { id: {{ $value->id }} })"
-                                                    >
-                                                        <i class="bx bx-trash me-1"></i> Delete
-                                                    </a>
-                
+                                                    @if(admin()->can('keeper-is-verify'))
+                                                        <a class="dropdown-item" href="javascript:void(0);"
+                                                            wire:click.prevent="$dispatch('keeperVerify', { id: {{ $value->id }} })"
+                                                        >
+                                                            <i class="bx bx-envelope me-1"></i> {{ $value->email_verified_at ? 'Not Verify' : 'Verify' }}
+                                                        </a>
+                                                    @endif
+
+                                                    @if(admin()->can('keeper-edit'))
+                                                        <a class="dropdown-item" href="javascript:void(0);"
+                                                            wire:click.prevent="$dispatch('keeperUpdate', { id: {{ $value->id }} })"
+                                                        >
+                                                            <i class="bx bx-edit-alt me-1"></i> Edit
+                                                        </a>
+                                                    @endif
+
+                                                    @if(admin()->can('keeper-delete'))
+                                                        <a class="dropdown-item" href="javascript:void(0);"
+                                                            wire:click.prevent="$dispatch('keeperDelete', { id: {{ $value->id }} })"
+                                                        >
+                                                            <i class="bx bx-trash me-1"></i> Delete
+                                                        </a>
+                                                    @endif
+
                                                 </div>
                                             </div>
                                         @else
