+++ resources/views/admin/pages/orders/partials/title.blade.php
@@ -10,6 +10,7 @@
             @endif
         </div>
         <div>
+            @if(admin()->can('transfer-print'))
             <a class="btn btn-sm btn-outline-primary" href="javascript:void(0);" onclick="openPrintReceipt({{ $order->id }})" ><i class='bx bx-printer'></i></a>
 
             <script type="text/javascript">
@@ -21,23 +22,29 @@
                     ).print();
                 }
             </script>
+            @endif
             @if($order->status == 'Open')
                 @if($order->type == 'Pending')
-                    <a class="btn btn-sm btn-outline-primary" href="javascript:void(0);"
-                        wire:click.prevent="$dispatch('orderUpdate', { id: {{ $order->id }} })"
-                    >
-                        <i class='bx bx-edit-alt'></i>
-                    </a>
+                    @if(admin()->can('transfer-edit'))
+                        <a class="btn btn-sm btn-outline-primary" href="javascript:void(0);"
+                            wire:click.prevent="$dispatch('orderUpdate', { id: {{ $order->id }} })"
+                        >
+                            <i class='bx bx-edit-alt'></i>
+                        </a>
+                        
+                        @livewire('admin.pages.orders.partials.edit', ['order' => $order], key('order-edit-'.time()))
+                    @endif
+
+                    @if(admin()->can('transfer-send'))
+                        {{-- @livewire('admin.pages.orders.partials.delete', ['order' => $order], key('order-delete-'.time())) --}}
+                        <a class="btn btn-sm btn-outline-primary" href="javascript:void(0);"
+                            wire:click.prevent="$dispatch('orderSend', { id: {{ $order->id }} })"
+                        >
+                            <i class='bx bx-send'></i>
+                        </a>
+                        @livewire('admin.pages.orders.partials.send', ['order' => $order], key('order-send-'.time()))
+                    @endif
                     
-                    @livewire('admin.pages.orders.partials.edit', ['order' => $order], key('order-edit-'.time()))
-                    
-                    {{-- @livewire('admin.pages.orders.partials.delete', ['order' => $order], key('order-delete-'.time())) --}}
-                    <a class="btn btn-sm btn-outline-primary" href="javascript:void(0);"
-                        wire:click.prevent="$dispatch('orderSend', { id: {{ $order->id }} })"
-                    >
-                        <i class='bx bx-send'></i>
-                    </a>
-                    @livewire('admin.pages.orders.partials.send', ['order' => $order], key('order-send-'.time()))
                 @endif
 
                     {{-- <a class="btn btn-sm btn-outline-primary" href="javascript:void(0);"
@@ -48,38 +55,48 @@
                     @livewire('admin.pages.orders.partials.convert-order', ['order' => $order], key('order-convert-'.time())) --}}
 
                 @if($order->type == 'Send')
-                    <a class="btn btn-sm btn-outline-primary" href="javascript:void(0);"
-                        wire:click.prevent="$dispatch('orderProcessed', { id: {{ $order->id }} })"
-                    >
-                        <i class='bx bx-loader-circle'></i>
-                    </a>
-                    @livewire('admin.pages.orders.partials.processed', ['order' => $order], key('order-processed-'.time()))
+                    @if(admin()->can('transfer-process'))
+                        <a class="btn btn-sm btn-outline-primary" href="javascript:void(0);"
+                            wire:click.prevent="$dispatch('orderProcessed', { id: {{ $order->id }} })"
+                        >
+                            <i class='bx bx-loader-circle'></i>
+                        </a>
+                        @livewire('admin.pages.orders.partials.processed', ['order' => $order], key('order-processed-'.time()))
+                    @endif
+                    
                 @endif
                 @if($order->type == 'Processed')
-                    <a class="btn btn-sm btn-outline-primary" href="javascript:void(0);"
-                        wire:click.prevent="$dispatch('orderShipped', { id: {{ $order->id }} })"
-                    >
-                        <i class='bx bxs-truck'></i>
-                    </a>
-                    @livewire('admin.pages.orders.partials.shipped', ['order' => $order], key('order-shipped-'.time()))
+                    @if(admin()->can('transfer-shipped'))
+
+                        <a class="btn btn-sm btn-outline-primary" href="javascript:void(0);"
+                            wire:click.prevent="$dispatch('orderShipped', { id: {{ $order->id }} })"
+                        >
+                            <i class='bx bxs-truck'></i>
+                        </a>
+                        @livewire('admin.pages.orders.partials.shipped', ['order' => $order], key('order-shipped-'.time()))
+                    @endif
                 @endif
 
                 @if($order->type == 'Shipped')
-                    <a class="btn btn-sm btn-outline-primary" href="javascript:void(0);"
-                        wire:click.prevent="$dispatch('orderSave', { id: {{ $order->id }} })"
-                    >
-                        <i class='bx bx-save'></i>
-                    </a>
-                    @livewire('admin.pages.orders.partials.save', ['order' => $order], key('order-save-'.time()))
+                    @if(admin()->can('transfer-save'))
+                        <a class="btn btn-sm btn-outline-primary" href="javascript:void(0);"
+                            wire:click.prevent="$dispatch('orderSave', { id: {{ $order->id }} })"
+                        >
+                            <i class='bx bx-save'></i>
+                        </a>
+                        @livewire('admin.pages.orders.partials.save', ['order' => $order], key('order-save-'.time()))
+                    @endif
                 @endif
             @endif
             @if($order->status == 'Open' && $order->type == 'Pending')
+                @if(admin()->can('transfer-delete'))
 
-                <a class="btn btn-sm btn-danger" href="javascript:void(0);"
-                    wire:click.prevent="$dispatch('orderDelete', { id: {{ $order->id }} })"
-                >
-                    <i class="bx bx-trash me-1"></i>
-                </a>
+                    <a class="btn btn-sm btn-danger" href="javascript:void(0);"
+                        wire:click.prevent="$dispatch('orderDelete', { id: {{ $order->id }} })"
+                    >
+                        <i class="bx bx-trash me-1"></i>
+                    </a>
+                @endif
             @endif
         </div>
     </div>
